\name{MRSP-internal}
\Rdversion{1.1}
\alias{Brier}
\alias{CUBbinomiallogit}
\alias{FPR.FNR}
\alias{FPR.FNR.MRSP.workhorse}
\alias{MRSP.control}                                 % im detail dokumentieren
\alias{MRSP.model}                                   % im detail dokumentieren
\alias{MRSPwelcomemessage}
\alias{MSE}
\alias{OLSreg}
\alias{Psplinetreshnonex}
\alias{Psplinetreshrefx}
\alias{Psplinetreshsymx}
\alias{SPG}
\alias{SPG.MRSP}
\alias{SPG.MRSP.cmpd}
\alias{SPGsmoothpen}
\alias{SPGsmoothpen.MRSP}
\alias{SPGsmoothpen.MRSP.cmpd}
\alias{bootstrap}                                    % im detail dokumentieren
\alias{colPsplinetreshnonex}
\alias{colPsplinetreshrefx}
\alias{colPsplinetreshsymx}
\alias{constrfun}
\alias{correctfun}
\alias{createResponse}
\alias{cv}                                           % im detail dokumentieren
\alias{evalh}
\alias{extract}
\alias{fista}
\alias{fista.control}
\alias{fistaProximal}
\alias{fistaProximal.MRSP}
\alias{fistaProximal.MRSP.cmpd}
\alias{getlambdamax}
\alias{getmlfit}
\alias{getpenweights}
\alias{l2norm}
\alias{l2norm.x}
\alias{l2norminv}
\alias{l2norminv.x}
\alias{l2normraw}
\alias{l2project}
\alias{lassoC}
\alias{lextract}
\alias{lmaxproject}
\alias{log1pexp}
\alias{log_row_sums}
\alias{lossapprox}
\alias{maxRow}
\alias{modelmatrix.MRSP}
\alias{multinom.simcoef}
\alias{multinom.simdata}
\alias{multinomlogit}                                   % als beispiel fuer MRSP.model angeben
\alias{n.knots}
\alias{penalty}
\alias{penalty.MRSP}
\alias{penalty.MRSP.cmpd}
\alias{proxfunGbridge}
\alias{proximterm}
\alias{pval}
\alias{refit}
\alias{rowCumprod}
\alias{rowCumsum}
\alias{rowFLtreshnonex}
\alias{rowL2norm}
\alias{rowL2norminv}
\alias{rowL2normraw}
\alias{se}
\alias{select}                                          % im detail dokumentieren
\alias{sequentiallogit}                                 % als beispiel fuer MRSP.model
\alias{simulation}
\alias{spFGLtreshnonex}
\alias{spFGLtreshsymx}
\alias{spGFLtreshnonex}
\alias{spectralnorm}
\alias{sptreshcs}
\alias{sptreshnonex}
\alias{sptreshrefx}
\alias{sptreshsymx}
\alias{summary.MRSP}
\alias{symfun}
\alias{treshcs}
\alias{treshg}
\alias{treshlcs}
\alias{treshlnonex}
\alias{treshlrefx}
\alias{treshlsymx}
\alias{treshmean}
\alias{treshnonex}
\alias{treshrefx}
\alias{treshsymx}
\alias{updateEta}
\alias{updatePenweights}
\alias{useSPG}
\alias{v.allequal}
\alias{which.max.cv}
\alias{which.min.cv}
\alias{yseqlog.constructor}
\alias{MRSP.bootstrap-class}
\alias{MRSP.coef-class}                                   % im detail dokumentieren
\alias{MRSP.control-class}                                % dito
\alias{MRSP.cv-class}                                     % ebenso
\alias{MRSP.model-class}                                  % auch
\alias{MRSP.se-class}
\alias{MRSP.sim-class}
\alias{fista.control-class}
\alias{Brier,matrix-method}
\alias{FPR.FNR,MRSP.sim-method}
\alias{MSE,MRSP.sim-method}
\alias{MSE,MRSP.coef-method}
\alias{SPG,MRSP.coef-method}
\alias{SPG,list-method}
\alias{SPGsmoothpen,MRSP.coef-method}
\alias{SPGsmoothpen,list-method}
\alias{createResponse,MRSP.coef-method}
\alias{fistaProximal,MRSP.coef-method}
\alias{lossapprox,ANY,ANY,MRSP.coef-method}
\alias{lossapprox,ANY,ANY,list-method}
\alias{penalty,MRSP.coef-method}
\alias{proximterm,MRSP.coef-method}
\alias{proximterm,list-method}
\alias{show,MRSP.coef-method}
\alias{simulation,MRSP.coef-method}
\alias{updateEta,ANY,MRSP.coef-method}
\alias{updatePenweights,ANY,MRSP.coef-method}
\alias{useSPG,MRSP.coef-method}
\alias{useSPG,list-method}
\alias{which.max.cv,numeric-method}
\alias{which.min.cv,numeric-method}

\title{Internal MRSP Functionality}
\description{
  Internal methods, functions and classes for package MRSP
}
\details{
 These functions are not be called by end-users.
}
\author{
Wolfgang Poessnecker
}
\keyword{internal}
